requirements.txt
torch>=2.2
torchvision>=0.17
flwr>=1.8
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
PyYAML>=6.0
tqdm>=4.66
scikit-learn>=1.4
opacus>=1.4
pytest>=8.0

